API

API

This page is a dump of all the docstrings found in the code.

CTModels.AbstractControlModel — Type

abstract type AbstractControlModel

CTModels.AbstractOptimalControlModel — Type

abstract type AbstractOptimalControlModel

CTModels.AbstractStateModel — Type

abstract type AbstractStateModel

CTModels.ControlModel — Type

struct ControlModel <: CTModels.AbstractControlModel

Fields

dimension::Int64

name::String

components::Vector{String}

CTModels.OptimalControlModel — Type

struct OptimalControlModel{ControlModelType<:CTModels.AbstractControlModel, StateModelType<:CTModels.AbstractStateModel} <: CTModels.AbstractOptimalControlModel

Fields

control::CTModels.AbstractControlModel

state::CTModels.AbstractStateModel

CTModels.OptimalControlModelMutable — Type

mutable struct OptimalControlModelMutable <: CTModels.AbstractOptimalControlModel

Fields

control::Union{Nothing, CTModels.AbstractControlModel}: Default: nothing

state::Union{Nothing, CTModels.AbstractStateModel}: Default: nothing

CTModels.StateModel — Type

struct StateModel <: CTModels.AbstractStateModel

Fields

dimension::Int64

name::String

components::Vector{String}

CTModels.__control_components — Method

__control_components(
    m::Int64,
    name::String
) -> Vector{String}

Used to set the default value of the names of the controls. The default value is ["u"] for a one dimensional control, and ["u₁", "u₂", ...] for a multi dimensional control.

CTModels.__control_name — Method

__control_name() -> String

Used to set the default value of the names of the control. The default value is "u".

CTModels.__is_control_set — Method

__is_control_set(
    ocp::CTModels.OptimalControlModelMutable
) -> Bool

CTModels.__is_state_set — Method

__is_state_set(
    ocp::CTModels.OptimalControlModelMutable
) -> Bool

CTModels.__state_components — Method

__state_components(n::Int64, name::String) -> Vector{String}

Used to set the default value of the names of the states. The default value is ["x"] for a one dimensional state, and ["x₁", "x₂", ...] for a multi dimensional state.

CTModels.__state_name — Method

__state_name() -> String

Used to set the default value of the name of the state. The default value is "x".

CTModels.control! — Method

control!(ocp::CTModels.OptimalControlModelMutable, m::Int64)
control!(
    ocp::CTModels.OptimalControlModelMutable,
    m::Int64,
    name::Union{String, Symbol}
)
control!(
    ocp::CTModels.OptimalControlModelMutable,
    m::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the control dimension and possibly the names of each coordinate.

Note

You must use control! only once to set the control dimension.

Examples

julia> control!(ocp, 1)
julia> control_dimension(ocp)
1
julia> control_components(ocp)
["u"]

julia> control!(ocp, 1, "v")
julia> control_dimension(ocp)
1
julia> control_components(ocp)
["v"]

julia> control!(ocp, 2)
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["u₁", "u₂"]

julia> control!(ocp, 2, :v)
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["v₁", "v₂"]

julia> control!(ocp, 2, "v")
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["v₁", "v₂"]

julia> control!(ocp, 2, "v", ["a", "b"])
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["a", "b"]

julia> control!(ocp, 2, "v", [:a, :b])
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["a", "b"]

CTModels.state! — Method

state!(ocp::CTModels.OptimalControlModelMutable, n::Int64)
state!(
    ocp::CTModels.OptimalControlModelMutable,
    n::Int64,
    name::Union{String, Symbol}
)
state!(
    ocp::CTModels.OptimalControlModelMutable,
    n::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the state dimension and possibly the names of each component.

Note

You must use state! only once to set the state dimension.

Examples

julia> state!(ocp, 1)
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["x"]

julia> state!(ocp, 1, "y")
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["y"]

julia> state!(ocp, 2)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["x₁", "x₂"]

julia> state!(ocp, 2, :y)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y")
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y", ["u", "v"])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]

julia> state!(ocp, 2, "y", [:u, :v])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]