Definition

Index

CTModels.definition

CTModels.definition

CTModels.definition!

Documentation

CTModels.definition! — Method

definition!(ocp::CTModels.PreModel, definition::Expr)

Set the model definition of the optimal control problem.

CTModels.definition — Method

definition(ocp::CTModels.Model) -> Expr

Return the model definition of the optimal control problem.

CTModels.definition — Method

definition(ocp::CTModels.PreModel) -> Union{Nothing, Expr}

Return the model definition of the optimal control problem or nothing.