State

Index

CTModels.components

CTModels.components

CTModels.dimension

CTModels.dimension

CTModels.name

CTModels.name

CTModels.state!

CTModels.value

Documentation

CTModels.components — Method

components(
    model::CTModels.StateModelSolution
) -> Vector{String}

Get the components names of the state from the state model solution.

CTModels.components — Method

components(model::CTModels.StateModel) -> Vector{String}

Get the components names of the state from the state model.

CTModels.dimension — Method

dimension(model::CTModels.StateModelSolution) -> Int64

Get the dimension of the state from the state model solution.

CTModels.dimension — Method

dimension(model::CTModels.StateModel) -> Int64

Get the dimension of the state from the state model.

CTModels.name — Method

name(model::CTModels.StateModelSolution) -> String

Get the name of the state from the state model solution.

CTModels.name — Method

name(model::CTModels.StateModel) -> String

Get the name of the state from the state model.

CTModels.state! — Method

state!(ocp::CTModels.PreModel, n::Int64)
state!(
    ocp::CTModels.PreModel,
    n::Int64,
    name::Union{String, Symbol}
)
state!(
    ocp::CTModels.PreModel,
    n::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the state dimension and possibly the names of each component.

Note

You must use state! only once to set the state dimension.

Examples

julia> state!(ocp, 1)
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["x"]

julia> state!(ocp, 1, "y")
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["y"]

julia> state!(ocp, 2)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["x₁", "x₂"]

julia> state!(ocp, 2, :y)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y")
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y", ["u", "v"])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]

julia> state!(ocp, 2, "y", [:u, :v])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]

CTModels.value — Method

value(
    model::CTModels.StateModelSolution{TS<:Function}
) -> Function

Get the state function from the state model solution.