Types

Index

CTModels.AbstractConstraintsModel

CTModels.AbstractControlModel

CTModels.AbstractDualModel

CTModels.AbstractModel

CTModels.AbstractObjectiveModel

CTModels.AbstractSolution

CTModels.AbstractSolverInfos

CTModels.AbstractStateModel

CTModels.AbstractTimeGridModel

CTModels.AbstractTimeModel

CTModels.AbstractTimesModel

CTModels.AbstractVariableModel

CTModels.Autonomous

CTModels.BolzaObjectiveModel

CTModels.ConstraintsModel

CTModels.ControlModel

CTModels.ControlModelSolution

CTModels.DualModel

CTModels.EmptyTimeGridModel

CTModels.EmptyVariableModel

CTModels.FixedTimeModel

CTModels.FreeTimeModel

CTModels.LagrangeObjectiveModel

CTModels.MayerObjectiveModel

CTModels.Model

CTModels.NonAutonomous

CTModels.PreModel

CTModels.Solution

CTModels.SolverInfos

CTModels.StateModel

CTModels.StateModelSolution

CTModels.TimeDependence

CTModels.TimeGridModel

CTModels.TimesModel

CTModels.VariableModel

CTModels.VariableModelSolution

CTModels.__is_autonomous_set

CTModels.__is_complete

CTModels.__is_consistent

CTModels.__is_control_set

CTModels.__is_control_set

CTModels.__is_definition_set

CTModels.__is_definition_set

CTModels.__is_dynamics_complete

CTModels.__is_dynamics_set

CTModels.__is_dynamics_set

CTModels.__is_empty

CTModels.__is_objective_set

CTModels.__is_objective_set

CTModels.__is_set

CTModels.__is_state_set

CTModels.__is_state_set

CTModels.__is_times_set

CTModels.__is_times_set

CTModels.__is_variable_empty

CTModels.__is_variable_set

CTModels.__is_variable_set

CTModels.is_empty_time_grid

CTModels.state_dimension

Documentation

CTModels.AbstractConstraintsModel — Type

abstract type AbstractConstraintsModel

CTModels.AbstractControlModel — Type

abstract type AbstractControlModel

CTModels.AbstractDualModel — Type

abstract type AbstractDualModel

CTModels.AbstractModel — Type

abstract type AbstractModel

CTModels.AbstractObjectiveModel — Type

abstract type AbstractObjectiveModel

CTModels.AbstractSolution — Type

abstract type AbstractSolution

CTModels.AbstractSolverInfos — Type

abstract type AbstractSolverInfos

CTModels.AbstractStateModel — Type

abstract type AbstractStateModel

CTModels.AbstractTimeGridModel — Type

abstract type AbstractTimeGridModel

CTModels.AbstractTimeModel — Type

abstract type AbstractTimeModel

CTModels.AbstractTimesModel — Type

abstract type AbstractTimesModel

CTModels.AbstractVariableModel — Type

abstract type AbstractVariableModel

CTModels.Autonomous — Type

abstract type Autonomous <: CTModels.TimeDependence

CTModels.BolzaObjectiveModel — Type

struct BolzaObjectiveModel{TM<:Function, TL<:Function} <: CTModels.AbstractObjectiveModel

Fields

mayer::Function

lagrange::Function

criterion::Symbol

CTModels.ConstraintsModel — Type

struct ConstraintsModel{TP<:Tuple, TB<:Tuple, TS<:Tuple, TC<:Tuple, TV<:Tuple} <: CTModels.AbstractConstraintsModel

Fields

path_nl::Tuple

boundary_nl::Tuple

state_box::Tuple

control_box::Tuple

variable_box::Tuple

CTModels.ControlModel — Type

struct ControlModel <: CTModels.AbstractControlModel

Fields

name::String

components::Vector{String}

CTModels.ControlModelSolution — Type

struct ControlModelSolution{TS<:Function} <: CTModels.AbstractControlModel

Fields

name::String

components::Vector{String}

value::Function

CTModels.DualModel — Type

struct DualModel{PC_Dual<:Union{Nothing, Function}, BC_Dual<:Union{Nothing, AbstractVector{<:Real}}, SC_LB_Dual<:Union{Nothing, Function}, SC_UB_Dual<:Union{Nothing, Function}, CC_LB_Dual<:Union{Nothing, Function}, CC_UB_Dual<:Union{Nothing, Function}, VC_LB_Dual<:Union{Nothing, AbstractVector{<:Real}}, VC_UB_Dual<:Union{Nothing, AbstractVector{<:Real}}} <: CTModels.AbstractDualModel

Fields

path_constraints_dual::Union{Nothing, Function}

boundary_constraints_dual::Union{Nothing, AbstractVector{<:Real}}

state_constraints_lb_dual::Union{Nothing, Function}

state_constraints_ub_dual::Union{Nothing, Function}

control_constraints_lb_dual::Union{Nothing, Function}

control_constraints_ub_dual::Union{Nothing, Function}

variable_constraints_lb_dual::Union{Nothing, AbstractVector{<:Real}}

variable_constraints_ub_dual::Union{Nothing, AbstractVector{<:Real}}

CTModels.EmptyTimeGridModel — Type

struct EmptyTimeGridModel <: CTModels.AbstractTimeGridModel

Fields

CTModels.EmptyVariableModel — Type

struct EmptyVariableModel <: CTModels.AbstractVariableModel

Fields

CTModels.FixedTimeModel — Type

struct FixedTimeModel{T<:Real} <: CTModels.AbstractTimeModel

Fields

time::Real

name::String

CTModels.FreeTimeModel — Type

struct FreeTimeModel <: CTModels.AbstractTimeModel

Fields

index::Int64

name::String

CTModels.LagrangeObjectiveModel — Type

struct LagrangeObjectiveModel{TL<:Function} <: CTModels.AbstractObjectiveModel

Fields

lagrange::Function

criterion::Symbol

CTModels.MayerObjectiveModel — Type

struct MayerObjectiveModel{TM<:Function} <: CTModels.AbstractObjectiveModel

Fields

mayer::Function

criterion::Symbol

CTModels.Model — Type

struct Model{TD<:CTModels.TimeDependence, TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, DynamicsModelType<:Function, ObjectiveModelType<:CTModels.AbstractObjectiveModel, ConstraintsModelType<:CTModels.AbstractConstraintsModel, BuildExaModelType<:Union{Nothing, Function}} <: CTModels.AbstractModel

Fields

times::CTModels.AbstractTimesModel

state::CTModels.AbstractStateModel

control::CTModels.AbstractControlModel

variable::CTModels.AbstractVariableModel

dynamics::Function

objective::CTModels.AbstractObjectiveModel

constraints::CTModels.AbstractConstraintsModel

definition::Expr

build_examodel::Union{Nothing, Function}

CTModels.NonAutonomous — Type

abstract type NonAutonomous <: CTModels.TimeDependence

CTModels.PreModel — Type

mutable struct PreModel <: CTModels.AbstractModel

Fields

times::Union{Nothing, CTModels.AbstractTimesModel}: Default: nothing

state::Union{Nothing, CTModels.AbstractStateModel}: Default: nothing

control::Union{Nothing, CTModels.AbstractControlModel}: Default: nothing

variable::CTModels.AbstractVariableModel: Default: EmptyVariableModel()

dynamics::Union{Nothing, Function, Vector{<:Tuple{var"#s52", var"#s53"} where {var"#s52"<:(AbstractRange{<:Int64}), var"#s53"<:Function}}}: Default: nothing

objective::Union{Nothing, CTModels.AbstractObjectiveModel}: Default: nothing

constraints::OrderedCollections.OrderedDict{Symbol, Tuple{Symbol, Union{Function, OrdinalRange{<:Int64}}, AbstractVector{<:Real}, AbstractVector{<:Real}}}: Default: ConstraintsDictType()

definition::Union{Nothing, Expr}: Default: nothing

autonomous::Union{Nothing, Bool}: Default: nothing

CTModels.Solution — Type

struct Solution{TimeGridModelType<:CTModels.AbstractTimeGridModel, TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, CostateModelType<:Function, ObjectiveValueType<:Real, DualModelType<:CTModels.AbstractDualModel, SolverInfosType<:CTModels.AbstractSolverInfos, ModelType<:CTModels.AbstractModel} <: CTModels.AbstractSolution

Fields

time_grid::CTModels.AbstractTimeGridModel

times::CTModels.AbstractTimesModel

state::CTModels.AbstractStateModel

control::CTModels.AbstractControlModel

variable::CTModels.AbstractVariableModel

costate::Function

objective::Real

dual::CTModels.AbstractDualModel

solver_infos::CTModels.AbstractSolverInfos

model::CTModels.AbstractModel

CTModels.SolverInfos — Type

struct SolverInfos{TI<:Dict{Symbol, Any}} <: CTModels.AbstractSolverInfos

Fields

iterations::Int64

status::Symbol

message::String

successful::Bool

constraints_violation::Float64

infos::Dict{Symbol, Any}

CTModels.StateModel — Type

struct StateModel <: CTModels.AbstractStateModel

Fields

name::String

components::Vector{String}

CTModels.StateModelSolution — Type

struct StateModelSolution{TS<:Function} <: CTModels.AbstractStateModel

Fields

name::String

components::Vector{String}

value::Function

CTModels.TimeDependence — Type

abstract type TimeDependence

CTModels.TimeGridModel — Type

struct TimeGridModel{T<:Union{StepRangeLen, AbstractVector{<:Real}}} <: CTModels.AbstractTimeGridModel

Fields

value::Union{StepRangeLen, AbstractVector{<:Real}}

CTModels.TimesModel — Type

struct TimesModel{TI<:CTModels.AbstractTimeModel, TF<:CTModels.AbstractTimeModel} <: CTModels.AbstractTimesModel

Fields

initial::CTModels.AbstractTimeModel

final::CTModels.AbstractTimeModel

time_name::String

CTModels.VariableModel — Type

struct VariableModel <: CTModels.AbstractVariableModel

Fields

name::String

components::Vector{String}

CTModels.VariableModelSolution — Type

struct VariableModelSolution{TS<:Union{Real, AbstractVector{<:Real}}} <: CTModels.AbstractVariableModel

Fields

name::String

components::Vector{String}

value::Union{Real, AbstractVector{<:Real}}

CTModels.__is_autonomous_set — Method

__is_autonomous_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_complete — Method

__is_complete(ocp::CTModels.PreModel) -> Bool

Return true if the PreModel can be built into a Model.

CTModels.__is_consistent — Method

__is_consistent(ocp::CTModels.PreModel) -> Bool

Return true if all the required fields are set in the PreModel.

CTModels.__is_control_set — Method

__is_control_set(ocp::CTModels.Model) -> Bool

CTModels.__is_control_set — Method

__is_control_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_definition_set — Method

__is_definition_set(ocp::CTModels.Model) -> Bool

CTModels.__is_definition_set — Method

__is_definition_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_dynamics_complete — Method

__is_dynamics_complete(ocp::CTModels.PreModel) -> Bool

CTModels.__is_dynamics_set — Method

__is_dynamics_set(ocp::CTModels.Model) -> Bool

CTModels.__is_dynamics_set — Method

__is_dynamics_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_empty — Method

__is_empty(ocp::CTModels.PreModel) -> Bool

Return true if nothing has been set.

CTModels.__is_objective_set — Method

__is_objective_set(ocp::CTModels.Model) -> Bool

CTModels.__is_objective_set — Method

__is_objective_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_set — Method

__is_set(x) -> Bool

CTModels.__is_state_set — Method

__is_state_set(ocp::CTModels.Model) -> Bool

CTModels.__is_state_set — Method

__is_state_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_times_set — Method

__is_times_set(ocp::CTModels.Model) -> Bool

CTModels.__is_times_set — Method

__is_times_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_variable_empty — Method

__is_variable_empty(v) -> Bool

CTModels.__is_variable_set — Method

__is_variable_set(ocp::CTModels.Model) -> Bool

CTModels.__is_variable_set — Method

__is_variable_set(ocp::CTModels.PreModel) -> Bool

CTModels.is_empty_time_grid — Method

is_empty_time_grid(sol::CTModels.Solution) -> Bool

Check if the time grid is empty from the solution.

CTModels.state_dimension — Method

state_dimension(ocp::CTModels.PreModel) -> Int64