API

API

This page is a dump of all the docstrings found in the code.

CTModels.AbstractControlModel — Type

abstract type AbstractControlModel

CTModels.AbstractModel — Type

abstract type AbstractModel

CTModels.AbstractObjectiveModel — Type

abstract type AbstractObjectiveModel

CTModels.AbstractStateModel — Type

abstract type AbstractStateModel

CTModels.AbstractTimeModel — Type

abstract type AbstractTimeModel

CTModels.AbstractTimesModel — Type

abstract type AbstractTimesModel

CTModels.AbstractVariableModel — Type

abstract type AbstractVariableModel

CTModels.BolzaObjectiveModel — Type

struct BolzaObjectiveModel{TM<:Function, TL<:Function} <: CTModels.AbstractObjectiveModel

Fields

mayer::Function

lagrange::Function

criterion::Symbol

CTModels.ControlModel — Type

struct ControlModel <: CTModels.AbstractControlModel

Fields

name::String

components::Vector{String}

CTModels.EmptyVariableModel — Type

struct EmptyVariableModel <: CTModels.AbstractVariableModel

Fields

CTModels.FixedTimeModel — Type

struct FixedTimeModel <: CTModels.AbstractTimeModel

Fields

time::Real

name::String

CTModels.FreeTimeModel — Type

struct FreeTimeModel <: CTModels.AbstractTimeModel

Fields

index::Int64

name::String

CTModels.LagrangeObjectiveModel — Type

struct LagrangeObjectiveModel{TL<:Function} <: CTModels.AbstractObjectiveModel

Fields

lagrange::Function

criterion::Symbol

CTModels.MayerObjectiveModel — Type

struct MayerObjectiveModel{TM<:Function} <: CTModels.AbstractObjectiveModel

Fields

mayer::Function

criterion::Symbol

CTModels.Model — Type

struct Model{TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, DynamicsModelType<:Function, ObjectiveModelType<:CTModels.AbstractObjectiveModel, ConstraintsModelType<:Dict{Symbol, Tuple{Symbol, Union{Function, OrdinalRange{<:Int64}}, AbstractVector{<:Real}, AbstractVector{<:Real}}}} <: CTModels.AbstractModel

Fields

times::CTModels.AbstractTimesModel

state::CTModels.AbstractStateModel

control::CTModels.AbstractControlModel

variable::CTModels.AbstractVariableModel

dynamics::Function

objective::CTModels.AbstractObjectiveModel

constraints::Dict{Symbol, Tuple{Symbol, Union{Function, OrdinalRange{<:Int64}}, AbstractVector{<:Real}, AbstractVector{<:Real}}}

definition::Expr

CTModels.PreModel — Type

mutable struct PreModel <: CTModels.AbstractModel

Fields

times::Union{Nothing, CTModels.AbstractTimesModel}: Default: nothing

state::Union{Nothing, CTModels.AbstractStateModel}: Default: nothing

control::Union{Nothing, CTModels.AbstractControlModel}: Default: nothing

variable::CTModels.AbstractVariableModel: Default: EmptyVariableModel()

dynamics::Union{Nothing, Function}: Default: nothing

objective::Union{Nothing, CTModels.AbstractObjectiveModel}: Default: nothing

constraints::Dict{Symbol, Tuple{Symbol, Union{Function, OrdinalRange{<:Int64}}, AbstractVector{<:Real}, AbstractVector{<:Real}}}: Default: ConstraintsDictType()

definition::Union{Nothing, Expr}: Default: nothing

CTModels.StateModel — Type

struct StateModel <: CTModels.AbstractStateModel

Fields

name::String

components::Vector{String}

CTModels.TimesModel — Type

struct TimesModel{TI<:CTModels.AbstractTimeModel, TF<:CTModels.AbstractTimeModel} <: CTModels.AbstractTimesModel

Fields

initial::CTModels.AbstractTimeModel

final::CTModels.AbstractTimeModel

time_name::String

CTModels.VariableModel — Type

struct VariableModel <: CTModels.AbstractVariableModel

Fields

name::String

components::Vector{String}

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    ocp::CTModels.Model
) -> Union{Nothing, Vector}

Print the optimal control problem.

CTModels.__constraint_label — Method

__constraint_label() -> Symbol

Used to set the default value of the label of a constraint. A unique value is given to each constraint using the gensym function and prefixing by :unamed.

CTModels.__control_components — Method

__control_components(
    m::Int64,
    name::String
) -> Vector{String}

Used to set the default value of the names of the controls. The default value is ["u"] for a one dimensional control, and ["u₁", "u₂", ...] for a multi dimensional control.

CTModels.__control_name — Method

__control_name() -> String

Used to set the default value of the names of the control. The default value is "u".

CTModels.__criterion_type — Method

__criterion_type() -> Symbol

Used to set the default value of the type of criterion. Either :min or :max. The default value is :min. The other possible criterion type is :max.

CTModels.__is_control_set — Method

__is_control_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_definition_set — Method

__is_definition_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_dynamics_set — Method

__is_dynamics_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_objective_set — Method

__is_objective_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_state_set — Method

__is_state_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_times_set — Method

__is_times_set(ocp::CTModels.PreModel) -> Bool

CTModels.__is_variable_set — Method

__is_variable_set(ocp::CTModels.PreModel) -> Bool

CTModels.__state_components — Method

__state_components(n::Int64, name::String) -> Vector{String}

Used to set the default value of the names of the states. The default value is ["x"] for a one dimensional state, and ["x₁", "x₂", ...] for a multi dimensional state.

CTModels.__state_name — Method

__state_name() -> String

Used to set the default value of the name of the state. The default value is "x".

CTModels.__time_name — Method

__time_name() -> String

Used to set the default value of the name of the time. The default value is t.

CTModels.__variable_components — Method

__variable_components(
    q::Int64,
    name::String
) -> Vector{String}

Used to set the default value of the names of the variables. The default value is ["v"] for a one dimensional variable, and ["v₁", "v₂", ...] for a multi dimensional variable.

CTModels.__variable_name — Method

__variable_name() -> String

Used to set the default value of the names of the variables. The default value is "v".

CTModels.control! — Method

control!(ocp::CTModels.PreModel, m::Int64)
control!(
    ocp::CTModels.PreModel,
    m::Int64,
    name::Union{String, Symbol}
)
control!(
    ocp::CTModels.PreModel,
    m::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the control dimension and possibly the names of each coordinate.

Note

You must use control! only once to set the control dimension.

Examples

julia> control!(ocp, 1)
julia> control_dimension(ocp)
1
julia> control_components(ocp)
["u"]

julia> control!(ocp, 1, "v")
julia> control_dimension(ocp)
1
julia> control_components(ocp)
["v"]

julia> control!(ocp, 2)
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["u₁", "u₂"]

julia> control!(ocp, 2, :v)
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["v₁", "v₂"]

julia> control!(ocp, 2, "v")
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["v₁", "v₂"]

julia> control!(ocp, 2, "v", ["a", "b"])
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["a", "b"]

julia> control!(ocp, 2, "v", [:a, :b])
julia> control_dimension(ocp)
2
julia> control_components(ocp)
["a", "b"]

CTModels.definition! — Method

definition!(ocp::CTModels.PreModel, definition::Expr)

Set the model definition of the optimal control problem.

CTModels.definition — Method

definition(ocp::CTModels.Model) -> Expr

Return the model definition of the optimal control problem or nothing.

CTModels.dim_boundary_cons_nl — Method

dim_boundary_cons_nl(ocp::CTModels.Model) -> Any

Return the dimension of the boundary constraints.

CTModels.dim_control_cons_box — Method

dim_control_cons_box(ocp::CTModels.Model) -> Any

Return the dimension of box constraints on control.

CTModels.dim_path_cons_nl — Method

dim_path_cons_nl(ocp::CTModels.Model) -> Any

Return the dimension of nonlinear path constraints.

CTModels.dim_state_cons_box — Method

dim_state_cons_box(ocp::CTModels.Model) -> Any

Return the dimension of box constraints on state.

CTModels.dim_variable_cons_box — Method

dim_variable_cons_box(ocp::CTModels.Model) -> Any

Return the dimension of box constraints on variable.

CTModels.dim_variable_cons_nl — Method

dim_variable_cons_nl(ocp::CTModels.Model) -> Any

Return the dimension of nonlinear variable constraints.

CTModels.objective! — Function

objective!(ocp::CTModels.PreModel; ...)
objective!(
    ocp::CTModels.PreModel,
    criterion::Symbol;
    mayer,
    lagrange
)

Set the objective of the optimal control problem.

Arguments

ocp::PreModel: the optimal control problem.

criterion::Symbol: the type of criterion. Either :min or :max. Default is :min.

mayer::Union{Function, Nothing}: the Mayer function (inplace). Default is nothing.

lagrange::Union{Function, Nothing}: the Lagrange function (inplace). Default is nothing.

Note

The state, control and variable must be set before the objective.

The objective must not be set before.

At least one of the two functions must be given. Please provide a Mayer or a Lagrange function.

Examples

```@example julia> function mayer!(r, x0, xf, v) r[1] = x0[1] + xf[1] + v[1] end juila> function lagrange!(r, t, x, u, v) r[1] = x[1] + u[1] + v[1] end julia> objective!(ocp, :min, mayer=mayer!, lagrange=lagrange!)

CTModels.state! — Method

state!(ocp::CTModels.PreModel, n::Int64)
state!(
    ocp::CTModels.PreModel,
    n::Int64,
    name::Union{String, Symbol}
)
state!(
    ocp::CTModels.PreModel,
    n::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the state dimension and possibly the names of each component.

Note

You must use state! only once to set the state dimension.

Examples

julia> state!(ocp, 1)
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["x"]

julia> state!(ocp, 1, "y")
julia> state_dimension(ocp)
1
julia> state_components(ocp)
["y"]

julia> state!(ocp, 2)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["x₁", "x₂"]

julia> state!(ocp, 2, :y)
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y")
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["y₁", "y₂"]

julia> state!(ocp, 2, "y", ["u", "v"])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]

julia> state!(ocp, 2, "y", [:u, :v])
julia> state_dimension(ocp)
2
julia> state_components(ocp)
["u", "v"]

CTModels.time! — Method

time!(ocp::CTModels.PreModel; t0, tf, ind0, indf, time_name)

Set the initial and final times. We denote by t0 the initial time and tf the final time. The optimal control problem is denoted ocp. When a time is free, then, one must provide the corresponding index of the ocp variable.

Note

You must use time! only once to set either the initial or the final time, or both.

Examples

julia> time!(ocp, t0=0,   tf=1  ) # Fixed t0 and fixed tf
julia> time!(ocp, t0=0,   indf=2) # Fixed t0 and free  tf
julia> time!(ocp, ind0=2, tf=1  ) # Free  t0 and fixed tf
julia> time!(ocp, ind0=2, indf=3) # Free  t0 and free  tf

When you plot a solution of an optimal control problem, the name of the time variable appears. By default, the name is "t". Consider you want to set the name of the time variable to "s".

julia> time!(ocp, t0=0, tf=1, name="s") # name is a String
# or
julia> time!(ocp, t0=0, tf=1, name=:s ) # name is a Symbol  

CTModels.variable! — Method

variable!(ocp::CTModels.PreModel, q::Int64)
variable!(
    ocp::CTModels.PreModel,
    q::Int64,
    name::Union{String, Symbol}
)
variable!(
    ocp::CTModels.PreModel,
    q::Int64,
    name::Union{String, Symbol},
    components_names::Array{T2<:Union{String, Symbol}, 1}
)

Define the variable dimension and possibly the names of each component.

Note

You can use variable! once to set the variable dimension.

Examples

julia> variable!(ocp, 1, "v")
julia> variable!(ocp, 2, "v", [ "v₁", "v₂" ])