Private API

This page lists non-exported (internal) symbols of CTModels.

From CTModels

final

CTModels.final — Function

final(
    model::CTModels.TimesModel{<:CTModels.AbstractTimeModel, TF<:CTModels.AbstractTimeModel}
) -> CTModels.AbstractTimeModel

Get the final time from the times model.

has_fixed_initial_time

CTModels.has_fixed_initial_time — Function

has_fixed_initial_time(
    times::CTModels.TimesModel{<:CTModels.FixedTimeModel{T<:Real}}
) -> Bool

Check if the initial time is fixed. Return true.

has_fixed_initial_time(
    times::CTModels.TimesModel{CTModels.FreeTimeModel}
) -> Bool

Check if the initial time is free. Return false.

has_fixed_initial_time(ocp::CTModels.Model) -> Bool

Check if the initial time is fixed.

index

CTModels.index — Function

index(model::CTModels.FreeTimeModel) -> Int64

Get the index of the time variable from the free time model.

initial

CTModels.initial — Function

initial(
    model::CTModels.TimesModel{TI<:CTModels.AbstractTimeModel}
) -> CTModels.AbstractTimeModel

Get the initial time from the times model.

initial_time

CTModels.initial_time — Function

initial_time(
    model::CTModels.TimesModel{<:CTModels.FixedTimeModel{T<:Real}}
) -> Real

Get the initial time from the times model, from a fixed initial time model.

initial_time(
    model::CTModels.TimesModel{CTModels.FreeTimeModel},
    variable::AbstractArray{T<:Real, 1}
) -> Any

Get the initial time from the times model, from a free initial time model.

initial_time(ocp::CTModels.AbstractModel) -> Real

Throw an error for unsupported initial time access.

initial_time(
    ocp::CTModels.AbstractModel,
    variable::AbstractVector
) -> Any

Throw an error for unsupported initial time access with variable.

initial_time(
    ocp::CTModels.Model{<:CTModels.TimeDependence, <:CTModels.TimesModel{CTModels.FixedTimeModel{T<:Real}}}
) -> Real

Return the initial time, for a fixed initial time.

initial_time(
    ocp::CTModels.Model{<:CTModels.TimeDependence, <:CTModels.TimesModel{CTModels.FreeTimeModel}},
    variable::AbstractArray{T<:Real, 1}
) -> Any

Return the initial time, for a free initial time.

initial_time(
    ocp::CTModels.Model{<:CTModels.TimeDependence, <:CTModels.TimesModel{CTModels.FreeTimeModel}},
    variable::Real
) -> Real

Return the initial time, for a free initial time.

is_final_time_fixed

CTModels.is_final_time_fixed — Function

Alias for has_fixed_final_time. Check if the final time is fixed.

Example

julia> is_final_time_fixed(times)  # equivalent to has_fixed_final_time(times)

See also: has_fixed_final_time, is_final_time_free.

is_final_time_free

CTModels.is_final_time_free — Function

Alias for has_free_final_time. Check if the final time is free.

Example

julia> is_final_time_free(times)  # equivalent to has_free_final_time(times)

See also: has_free_final_time, is_final_time_fixed.

is_initial_time_fixed

CTModels.is_initial_time_fixed — Function

Alias for has_fixed_initial_time. Check if the initial time is fixed.

Example

julia> is_initial_time_fixed(times)  # equivalent to has_fixed_initial_time(times)

See also: has_fixed_initial_time, is_initial_time_free.

is_initial_time_free

CTModels.is_initial_time_free — Function

Alias for has_free_initial_time. Check if the initial time is free.

Example

julia> is_initial_time_free(times)  # equivalent to has_free_initial_time(times)

See also: has_free_initial_time, is_initial_time_fixed.

name

CTModels.name — Function

name(model::CTModels.StateModel) -> String

Get the name of the state from the state model.

name(model::CTModels.StateModelSolution) -> String

Get the name of the state from the state model solution.

name(model::CTModels.ControlModel) -> String

Get the name of the control variable.

Arguments

model::ControlModel: The control model.

Returns

String: The name of the control.

Example

julia> name(controlmodel)
"u"

name(model::CTModels.ControlModelSolution) -> String

Get the name of the control variable from the solution.

Arguments

model::ControlModelSolution: The control model solution.

Returns

String: The name of the control.

name(model::CTModels.VariableModel) -> String

Return the name of the variable stored in the model.

name(model::CTModels.VariableModelSolution) -> String

Return the name of the variable stored in the model solution.

name(_::CTModels.EmptyVariableModel) -> String

Return an empty string, since no variable is defined.

name(model::CTModels.FixedTimeModel) -> String

Get the name of the time from the fixed time model.

name(model::CTModels.FreeTimeModel) -> String

Get the name of the time from the free time model.

time

CTModels.time — Function

time(model::CTModels.FixedTimeModel{T<:Real}) -> Real

Get the time from the fixed time model.

time(
    model::CTModels.FreeTimeModel,
    variable::AbstractArray{T<:Real, 1}
) -> Any

Get the time from the free time model.

Exceptions

If the index of the time variable is not in [1, length(variable)], throw an error.

time!

CTModels.time! — Function

time!(ocp::CTModels.PreModel; t0, tf, ind0, indf, time_name)

Set the initial and final times. We denote by t0 the initial time and tf the final time. The optimal control problem is denoted ocp. When a time is free, then, one must provide the corresponding index of the ocp variable.

Note

You must use time! only once to set either the initial or the final time, or both.

Examples

julia> time!(ocp, t0=0,   tf=1  ) # Fixed t0 and fixed tf
julia> time!(ocp, t0=0,   indf=2) # Fixed t0 and free  tf
julia> time!(ocp, ind0=2, tf=1  ) # Free  t0 and fixed tf
julia> time!(ocp, ind0=2, indf=3) # Free  t0 and free  tf

When you plot a solution of an optimal control problem, the name of the time variable appears. By default, the name is "t". Consider you want to set the name of the time variable to "s".

julia> time!(ocp, t0=0, tf=1, time_name="s") # time_name is a String
# or
julia> time!(ocp, t0=0, tf=1, time_name=:s ) # time_name is a Symbol